Added our entire AT&T Shape Hackethon project
-Added our entire AT&T Shape Hackethon project
for 'FindMyFlix' a .NET MVC 5 application that utilizes AngularJS on the
front-end and .NET MVC on the backend. The application utilizes IBM
Watson's Speech-To-Text library, Conversation, and Text-To-Speech, as
well as data from Warner Brother's who provided their movie data from
the flixster library. The application is used for letting the user speak
to the application with a request for a movie genre(i.e. "I want to
watch something funny."), the IBM Watson API then turns this speech into
an intent which grabs the comedy data to be displayed for the user. As
the movies for the genre are displayed, the text-to-speech library of
IBM Watson then translates a statement for the user(i.e. IBm Watson:
"These are the movies within the comedy category for you.")